requirements.txt
replit==3.2.15
torch==2.4.0
transformers==4.40.0
coqui-tts==0.9.0
whisper-micro==0.5
soundfile==0.12.1
